Move code to "internal" package.
