+++ services/regulated-assets-approval-server/cmd/configure-asset-issuer.go
@@ -6,7 +6,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stellar/go/clients/horizonclient"
 	"github.com/stellar/go/network"
-	"github.com/stellar/go/services/regulated-assets-approval-server/configureassetissuer"
+	"github.com/stellar/go/services/regulated-assets-approval-server/internal/configureassetissuer"
 	"github.com/stellar/go/support/config"
 )
